\name{findMode}
\alias{findMode}

\title{Find Mode for a Given Sample}

\usage{
findMode(x)
}

\arguments{
  \item{x}{a vector of samples.}
}

\description{
  This function finds the mode of empirical density functior for given sample.}

\details{
  ...
} 
\value{
 The value of mode.
}

\author{
  Liang Jing \email{ljing918@gmail.com}
  }

\examples{
\dontrun{
findMode(rnorm(1000))
}
}

\seealso{

}

\keyword{mode}
