\name{TexasCounty.boundary}
\alias{TexasCounty.boundary}

\title{Data Set of Texas County Boundries}

\usage{
data(TexasCounty_boundary)
}

\arguments{

}

\description{
  This data set contains the boundary information for all Texas countries.}

\details{
  ...
} 
\value{

}

\author{
  Liang Jing \email{ljing918@gmail.com}
  }

\examples{
\dontrun{

data(TexasCounty_boundary)
length(TexasCounty.boundary)
plotDataBD(TexasCounty.boundary)
}
}

\seealso{
\code{\link{plotDataBD}}, \code{\link{TexasCounty.center}}, \code{\link{TexasCounty.population}}.
}

\keyword{Data}
