\name{TexasCounty.population}
\alias{TexasCounty.population}

\title{Data Set of Texas County Population}

\usage{
data(TexasCounty_population)
}

\arguments{

}

\description{
  This data set contains the population information for all Texas countries.}

\details{
  ...
} 
\value{

}

\author{
  Liang Jing \email{ljing918@gmail.com}
  }

\examples{
\dontrun{

data(TexasCounty_population)
str(TexasCounty.population)
}
}

\seealso{
\code{\link{TexasCounty.boundary}}, \code{\link{TexasCounty.center}}.
}

\keyword{Data}
