\name{Weed}
\alias{Weed}

\title{Data Set of Weed}

\usage{
data(datWeed)
}

\arguments{

}

\description{
  This data set contains the Weed data.}

\details{
  ...
} 
\value{

}

\author{
  Liang Jing \email{ljing918@gmail.com}
  }

\examples{
\dontrun{

data(datWeed)
str(Weed)
plotData(Weed[,3], Weed[,1:2], Weed[,4], Weed[,1:2])
}
}

\seealso{
\code{\link{plotData}}.
}

\keyword{Data}
