\name{TexasCounty.population}
\alias{TexasCounty.population}

\title{Data Set of Texas County Population}

\usage{
data(TexasCounty_population)
}

\arguments{

}

\description{
  This data set contains the population information for all Texas countries.}

\details{
   Year: 2009
   
   Source: U.S. Census Bureau, Small Area Estimates Branch, Poverty and Median Income Estimates   
} 
\value{
  A data.frame with 254 observations and 3 variables.
}

\author{
  Liang Jing \email{ljing918@gmail.com}
  }

\examples{
\dontrun{
data(TexasCounty_population)
str(TexasCounty.population)
}
}

\seealso{
\code{\link{TexasCounty.boundary}}, \code{\link{TexasCounty.center}}.
}

\keyword{Data}
